refactor(elreal): consolidate exact-dyadic oracle helpers into one shared header (#1035)
The exact-dyadic oracle helpers (exact_real / exact_block / exact_blocks /
exact_value / exact_series_sum / approx / check_zero_overlap), built on
dyadic_exact.hpp (the #1022 oracle work), had been copy-pasted into FIVE places
that had already drifted apart: addition.cpp and add_renormalization.cpp
(inline), exact_value_oracle.cpp (as exact_zbcl), summation/summation_oracle.hpp
(z.take(32)), and arithmetic/arithmetic_oracle.hpp (z.take(64), added after the
issue with a "consolidation tracked by #1035" note). Keeping the subtle
digits<=53 vs wide-cfloat bit path in sync by hand was error-prone, and the
32-vs-64 window split was exactly the drift #1035 warned about.

Promote them to a single source of truth:

  include/sw/universal/verification/elreal_oracle.hpp
  namespace sw::universal::elreal_oracle

All helpers are templates (implicitly inline), so multi-TU inclusion is ODR-safe.
The window is unified to ZBCL_EXACT_WINDOW = 32 (the #1022 value). This is
behaviour-preserving: a ZBCL<double> near 1.0 saturates at ~19-20 non-overlapping
components against the 2^-1022 floor, so take(32) == take(64) for every test
value; the arithmetic suite's former 64-window captured nothing extra.

- New: verification/elreal_oracle.hpp (superset of all five copies).
- Deleted: summation/summation_oracle.hpp, arithmetic/arithmetic_oracle.hpp.
- 17 elreal test TUs now include the shared header; the `namespace est = ...`
  aliases retarget to elreal_oracle (call sites unchanged). addition.cpp,
  add_renormalization.cpp, and exact_value_oracle.cpp drop their inline copies
  for a using-directive; exact_value_oracle.cpp's exact_zbcl call sites become
  exact_value. exact_value_oracle.cpp keeps its EFT-specific oracle code
  (dval / block_ulp / EFT checks) local.

Pure refactor, no behaviour change. Verified: el_arith_*, el_sum_*, el_math_*
(20 targets) build and pass with BOTH gcc and clang. ASCII clean.

Resolves #1035
