+++ elastic/elreal/summation/exp_preview.cpp
@@ -22,11 +22,11 @@
 #include <universal/verification/dyadic_exact.hpp>
 #include <universal/verification/test_suite.hpp>
 
-#include "summation_oracle.hpp"
+#include <universal/verification/elreal_oracle.hpp>
 
 namespace {
 
-namespace est = sw::universal::elreal_sum_test;
+namespace est = sw::universal::elreal_oracle;
 
 // Taylor term n of exp(1): 1/n!.
 template <typename FpType>
